pyproject.toml
[project]
name = "feed"
version = "0.1.0"
description = "YouTube data aggregation and analysis system"
requires-python = ">=3.13"
dependencies = [
    "deepdiff",
    "google-api-python-client",
    "google-auth-oauthlib",
    "isodate",
    "langchain-core",
    "langchain-openai",
    "youtube-transcript-api",
]

[project.optional-dependencies]
web = [
    "browser-cookie3",
    "playwright",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.9",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["lib"]

[tool.ruff]
target-version = "py313"
src = ["lib"]

[tool.ruff.lint]
select = ["F", "E", "W", "I"]
ignore = ["E402", "E501", "E701", "E711", "E722", "E741", "F401", "F821", "F841", "W291", "W293"]

[tool.ruff.lint.isort]
known-local-folder = ["youtube", "analysis", "config", "context", "util"]
